Move util methods to util.py
